backend/requirements.txt


# fastapi==0.103.2
# uvicorn==0.23.2

# asyncpg==0.29.0
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# passlib==1.7.4
# bcrypt==4.0.1
# python-jose[cryptography]==3.3.0
# python-dotenv==1.0.0
# python-multipart==0.0.9

# aiofiles==23.2.1
# pydantic==1.10.26

# requests
# beautifulsoup4
# newspaper3k
# warcio

# pandas
# openpyxl
# xlrd==1.2.0
# python-docx
# python-pptx
# pdfplumber
# pillow

# scikit-learn
# nltk
# torch
# transformers

# google-genai

# six











# --- Core API ---
fastapi>=0.110
starlette>=0.36
uvicorn>=0.27

# --- Async / Validation ---
pydantic>=2.6
anyio>=4

# --- Database ---
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# --- Auth / Security ---
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.9

# --- File handling ---
aiofiles==23.2.1
pillow
pdfplumber
python-docx
python-pptx
pandas
openpyxl
xlrd>=2.0.1

# --- Scraping / Web ---
requests
beautifulsoup4
newspaper3k
warcio

# --- ML / NLP ---
scikit-learn
nltk
torch
transformers

# --- Gemini OCR ---
google-genai

slowapi